refactor: moved to separate folder "modals"
